Fix: point the stack-name conflict check at the moved Stack type
The name-conflict check landed on main while this branch was moving
shared modules by import direction, so it imports `pages/stacks/types`,
which is now `api/stack-types`. Nothing git could see conflicted; the
build is what caught it.
